Rearrange files structure and directory
