--- app/src/main/java/tarsila/costalonga/notasapp/presenter/BaseApplication.kt
@@ -1,11 +1,0 @@
-package tarsila.costalonga.notasapp.presenter
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class BaseApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
