--- app/src/main/java/tarsila/costalonga/notasapp/presenter/detalhefragment/EnumHelper.kt
@@ -1,11 +1,0 @@
-package tarsila.costalonga.notasapp.presenter.detalhefragment
-
-enum class MenuType {
-    SHARE,
-    DELETE,
-}
-
-enum class DetailMode {
-    VIEW,
-    EDIT,
-}
