--- app/src/main/java/tarsila/costalonga/notasapp/presenter/estatisticasfragment/EstatisticasFragment.kt
@@ -1,34 +1,0 @@
-package tarsila.costalonga.notasapp.presenter.estatisticasfragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import dagger.hilt.android.AndroidEntryPoint
-import tarsila.costalonga.notasapp.compose.EstatisticasScreen
-import tarsila.costalonga.notasapp.compose.theme.NotaComposeTheme
-
-@AndroidEntryPoint
-class EstatisticasFragment : Fragment() {
-    private val viewModel: EstatisticasViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        viewModel.carregarNotas()
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                NotaComposeTheme {
-                    EstatisticasScreen(viewModel)
-                }
-            }
-        }
-    }
-}
