--- app/src/main/java/tarsila/costalonga/notasapp/presenter/sobreFragment/SobreFragment.kt
@@ -1,28 +1,0 @@
-package tarsila.costalonga.notasapp.presenter.sobreFragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import tarsila.costalonga.notasapp.compose.SobreDescription
-import tarsila.costalonga.notasapp.compose.theme.NotaComposeTheme
-
-class SobreFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                NotaComposeTheme {
-                    SobreDescription()
-                }
-            }
-        }
-    }
-}
