--- app/src/main/java/tarsila/costalonga/notasapp/presenter/utils/UtilsFunctions.kt
@@ -1,11 +1,0 @@
-package tarsila.costalonga.notasapp.presenter.utils
-
-import android.content.Context
-import android.widget.Toast
-
-fun makeToast(
-    context: Context,
-    msg: String,
-) {
-    Toast.makeText(context, msg, Toast.LENGTH_SHORT).show()
-}
