+++ app/src/main/java/tarsila/costalonga/notasapp/ui/BaseApplication.kt
@@ -1,0 +1,11 @@
+package tarsila.costalonga.notasapp.ui
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class BaseApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
