+++ app/src/main/java/tarsila/costalonga/notasapp/ui/MainActivity.kt
@@ -1,0 +1,35 @@
+package tarsila.costalonga.notasapp.ui
+
+import android.content.res.Configuration
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.core.view.WindowCompat
+import dagger.hilt.android.AndroidEntryPoint
+import tarsila.costalonga.notasapp.databinding.ActivityMainBinding
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        installSplashScreen()
+        configureStatusBarIconsColor()
+
+        setContentView(binding.root)
+    }
+
+    private fun configureStatusBarIconsColor() {
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+        val controller = WindowCompat.getInsetsController(window, window.decorView)
+        controller.let { insetsController ->
+            insetsController.isAppearanceLightStatusBars = isSystemInLightTheme()
+        }
+    }
+
+    private fun AppCompatActivity.isSystemInLightTheme(): Boolean {
+        return resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK != Configuration.UI_MODE_NIGHT_YES
+    }
+}
