+++ app/src/main/java/tarsila/costalonga/notasapp/ui/compose/util/TextExt.kt
@@ -1,0 +1,13 @@
+package tarsila.costalonga.notasapp.ui.compose.util
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.text.style.TextDecoration
+
+@Composable
+fun getTextDecoration(finalizado: Boolean): TextDecoration {
+    return if (finalizado) {
+        TextDecoration.LineThrough
+    } else {
+        TextDecoration.None
+    }
+}
