+++ app/src/main/java/tarsila/costalonga/notasapp/ui/detailnote/EnumHelper.kt
@@ -1,0 +1,11 @@
+package tarsila.costalonga.notasapp.ui.detailnote
+
+enum class MenuType {
+    SHARE,
+    DELETE,
+}
+
+enum class DetailMode {
+    VIEW,
+    EDIT,
+}
