+++ app/src/main/res/layout/sobre_text.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".presenter.sobreFragment.SobreFragment">
+    tools:context=".ui.aboutsection.SobreFragment">
 
     <TextView
         android:id="@+id/txt_conteudo"
